SCRUM-185 tests: refactored IT tests - separated full valid flow from other order update tests
